refactor: 게시판 관련 리팩토링
