+++ app/src/main/java/com/kau/ttokttok/_core/domain/DomainModule.kt
@@ -1,11 +1,13 @@
 package com.kau.ttokttok._core.domain
 
 import com.kau.ttokttok.data.local.repository.CommunityRepositoryImpl
+import com.kau.ttokttok.data.local.repository.NoiseVoteRepositoryImpl
 import com.kau.ttokttok.data.local.repository.ReportRepositoryImpl
 import com.kau.ttokttok.domain.usecase.community.CreatePostUseCase
 import com.kau.ttokttok.domain.usecase.community.GetPostDetailUseCase
 import com.kau.ttokttok.domain.usecase.community.LoadPostsUseCase
 import com.kau.ttokttok.domain.usecase.monthreport.GetMonthReportUseCase
+import com.kau.ttokttok.domain.usecase.noisevote.LoadPostsNoiseVoteUseCase
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -14,6 +16,13 @@ import dagger.hilt.components.SingletonComponent
 @Module
 @InstallIn(SingletonComponent::class)
 class DomainModule {
+    // 소음 현황판
+    @Provides
+    fun provideLoadPostsNoiseVoteUseCase(
+        repository: NoiseVoteRepositoryImpl
+    ): LoadPostsNoiseVoteUseCase {
+        return LoadPostsNoiseVoteUseCase(repository)
+    }
 
     // 공지 사항
     @Provides
