+++ app/src/main/java/com/kau/ttokttok/data/remote/dto/noiseboard/req/FirstVoteReq.kt
@@ -1,6 +1,6 @@
 package com.kau.ttokttok.data.remote.dto.noiseboard.req
 
-import com.kau.ttokttok.domain.model.board.VoteType
+import com.kau.ttokttok.domain.model.board.noisevote.VoteType
 
 data class FirstVoteReq(
     val voteType: VoteType
