+++ app/src/main/java/com/kau/ttokttok/data/remote/dto/noiseboard/res/FirstVoteRes.kt
@@ -1,6 +1,6 @@
 package com.kau.ttokttok.data.remote.dto.noiseboard.res
 
-import com.kau.ttokttok.domain.model.board.VoteType
+import com.kau.ttokttok.domain.model.board.noisevote.VoteType
 import java.time.LocalDateTime
 
 data class FirstVoteRes(
