+++ app/src/main/java/com/kau/ttokttok/data/remote/dto/noiseboard/res/GetPostsNoiseBoardRes.kt
@@ -1,6 +1,9 @@
 package com.kau.ttokttok.data.remote.dto.noiseboard.res
 
+import android.health.connect.LocalTimeRangeFilter
+import com.kau.ttokttok.domain.model.board.noisevote.NoiseVoteBoard
 import java.time.LocalDateTime
+import java.time.format.DateTimeFormatter
 
 data class GetPostsNoiseBoardRes(
     val reports: List<NoiseBoard>,
@@ -17,5 +20,15 @@ data class NoiseBoard(
     val authorDong: Int,
     val reportedAt: String,
     val category: String,
-    val summary: String
+    val summary: String?
 )
+
+fun NoiseBoard.toNoiseVoteBoard(): NoiseVoteBoard {
+    return NoiseVoteBoard(
+        id = reportId,
+        authorLocation = String.format("%s동", authorDong),
+        title = summary ?: "",
+        reportedAt = LocalDateTime.parse(reportedAt, DateTimeFormatter.ISO_DATE_TIME),
+        category = category
+    )
+}
