+++ app/src/main/java/com/kau/ttokttok/data/remote/dto/preconsideration/res/GetPostDetailPreConsiderationRes.kt
@@ -1,11 +1,8 @@
 package com.kau.ttokttok.data.remote.dto.preconsideration.res
 
-import java.time.LocalDateTime
-
 data class GetPostDetailPreConsiderationRes(
     val preNoticeId: Long,
     val authorDong: Int,
-    val authorHosu: Int,
     val title: String,
     val content: String,
     val eventDate: String,
