--- app/src/main/java/com/kau/ttokttok/domain/exception/AppException.kt
@@ -1,5 +1,0 @@
-package com.kau.ttokttok.domain.exception
-
-sealed class AppException : Exception() {
-
-}
