+++ app/src/main/java/com/kau/ttokttok/ui/compose/community/detail/CommunityDetailViewModel.kt
@@ -19,8 +19,7 @@ data class CommunityDetailUiState(
     val communityBoardDetail: CommunityBoardDetail = CommunityBoardDetail(
         title = "제목 NULL",
         createdAt = LocalDateTime.now(),
-        content = "내용 NULL",
-        author = "관리자"
+        content = "내용 NULL"
     )
 )
