+++ app/src/main/java/com/kau/ttokttok/ui/compose/noisevote/NoiseVoteRoute.kt
@@ -10,10 +10,10 @@ fun NoiseVoteRoute(
     viewModel: NoiseVoteViewModel = hiltViewModel(),
     onClickPost: (Long) -> Unit
 ) {
-    val posts by viewModel.posts.collectAsStateWithLifecycle()
+    val uiState by viewModel.uiState.collectAsStateWithLifecycle()
 
     NoiseVoteScreen(
-        onClickPost = onClickPost,
-        posts = posts
+        uiState = uiState,
+        onClickPost = onClickPost
     )
 }
