+++ app/src/main/java/com/kau/ttokttok/ui/theme/Color.kt
@@ -36,3 +36,10 @@ val Purple900 = Color(0xFF4C1D95)     // 진함
 // ---------------------
 val Red400 = Color(0xFFF87171)
 val Green400 = Color(0xFF34D399)
+
+val Orange50  = Color(0xFFFFF7ED)
+val Orange200 = Color(0xFFFECBA1)
+val Orange900 = Color(0xFF7C2D12)
+val Gray700   = Color(0xFF374151)
+val Gray900   = Color(0xFF111827)
+val Gray50Bg  = Color(0xFFF9FAFB)
